fix: move TextEditorButton to a separate component
