+++ packages/picasso-lab/src/TextEditorButton/TextEditorButton.tsx
@@ -4,15 +4,17 @@ import { BaseProps, Button } from '@toptal/picasso'
 import { makeStyles, Theme } from '@material-ui/core'
 
 import styles from './styles'
-import { ButtonHandlerType } from './types'
 
 type Props = BaseProps & {
   icon: ReactElement
-  onClick?: ButtonHandlerType
+  onClick?: (e: React.MouseEvent<HTMLButtonElement>) => void
   active?: boolean
   disabled?: boolean
 }
 
+// Using { index: 10 } to inject CSS generated classes after the button's classes
+// in order to prevent Button's styles to override custom TextEditorButton styles
+// Related Jira issue: https://toptal-core.atlassian.net/browse/FX-1520
 const useStyles = makeStyles<Theme>(styles, {
   name: 'TextEditorButton',
   index: 10
