+++ packages/picasso-lab/src/TextEditorButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TextEditorButton'
